Moved sweeper func to sweep.go
